Moved Client, ClientSend and ClientReceive back to gui package
